Move ConnectionData to models package
