+++ stream-chat-android-client/api/stream-chat-android-client.api
@@ -2036,6 +2036,19 @@ public final class io/getstream/chat/android/client/models/Config {
 	public fun toString ()Ljava/lang/String;
 }
 
+public final class io/getstream/chat/android/client/models/ConnectionData {
+	public fun <init> (Lio/getstream/chat/android/client/models/User;Ljava/lang/String;)V
+	public final fun component1 ()Lio/getstream/chat/android/client/models/User;
+	public final fun component2 ()Ljava/lang/String;
+	public final fun copy (Lio/getstream/chat/android/client/models/User;Ljava/lang/String;)Lio/getstream/chat/android/client/models/ConnectionData;
+	public static synthetic fun copy$default (Lio/getstream/chat/android/client/models/ConnectionData;Lio/getstream/chat/android/client/models/User;Ljava/lang/String;ILjava/lang/Object;)Lio/getstream/chat/android/client/models/ConnectionData;
+	public fun equals (Ljava/lang/Object;)Z
+	public final fun getConnectionId ()Ljava/lang/String;
+	public final fun getUser ()Lio/getstream/chat/android/client/models/User;
+	public fun hashCode ()I
+	public fun toString ()Ljava/lang/String;
+}
+
 public final class io/getstream/chat/android/client/models/ContentUtils {
 	public static final fun getImage (Lio/getstream/chat/android/client/models/Channel;)Ljava/lang/String;
 	public static final fun getImage (Lio/getstream/chat/android/client/models/User;)Ljava/lang/String;
@@ -2580,19 +2593,6 @@ public final class io/getstream/chat/android/client/notifications/handler/Notifi
 	public final fun getNotificationChannelName ()I
 	public final fun getSmallIcon ()I
 	public final fun getUseProvidedFirebaseInstance ()Z
-	public fun hashCode ()I
-	public fun toString ()Ljava/lang/String;
-}
-
-public final class io/getstream/chat/android/client/socket/ConnectionData {
-	public fun <init> (Lio/getstream/chat/android/client/models/User;Ljava/lang/String;)V
-	public final fun component1 ()Lio/getstream/chat/android/client/models/User;
-	public final fun component2 ()Ljava/lang/String;
-	public final fun copy (Lio/getstream/chat/android/client/models/User;Ljava/lang/String;)Lio/getstream/chat/android/client/socket/ConnectionData;
-	public static synthetic fun copy$default (Lio/getstream/chat/android/client/socket/ConnectionData;Lio/getstream/chat/android/client/models/User;Ljava/lang/String;ILjava/lang/Object;)Lio/getstream/chat/android/client/socket/ConnectionData;
-	public fun equals (Ljava/lang/Object;)Z
-	public final fun getConnectionId ()Ljava/lang/String;
-	public final fun getUser ()Lio/getstream/chat/android/client/models/User;
 	public fun hashCode ()I
 	public fun toString ()Ljava/lang/String;
 }
