+++ stream-chat-android-client/src/androidTest/java/io/getstream/chat/android/client/utils/TestInitCallback.kt
@@ -2,7 +2,7 @@ package io.getstream.chat.android.client.utils
 
 import io.getstream.chat.android.client.call.Call
 import io.getstream.chat.android.client.errors.ChatError
-import io.getstream.chat.android.client.socket.ConnectionData
+import io.getstream.chat.android.client.models.ConnectionData
 
 internal class TestInitCallback : Call.Callback<ConnectionData> {
     private var data: ConnectionData? = null
