+++ stream-chat-android-client/src/main/java/io/getstream/chat/android/client/models/ConnectionData.kt
@@ -1,0 +1,3 @@
+package io.getstream.chat.android.client.models
+
+public data class ConnectionData(val user: User, val connectionId: String)
